Higher Level MFA and Recovery

+++ app/controllers/multifactor_auths_controller.rb
@@ -1,8 +1,14 @@
 class MultifactorAuthsController < ApplicationController
+  include MfaExpiryMethods
+  include WebauthnVerifiable
+
   before_action :redirect_to_signin, unless: :signed_in?
-  before_action :require_mfa_disabled, only: %i[new create]
+  before_action :require_totp_disabled, only: %i[new create]
   before_action :require_mfa_enabled, only: :update
+  before_action :require_totp_enabled, only: %i[mfa_update destroy]
   before_action :seed_and_expire, only: :create
+  before_action :verify_session_expiration, only: %i[mfa_update webauthn_update]
+  after_action :delete_mfa_level_update_session_variables, only: %i[mfa_update webauthn_update]
   helper_method :issuer
 
   def new
@@ -21,20 +27,73 @@ class MultifactorAuthsController < ApplicationController
     else
       flash.now[:success] = t(".success")
       @continue_path = session.fetch("mfa_redirect_uri", edit_settings_path)
-      session.delete("mfa_redirect_uri")
-      render :recovery
+
+      if current_user.mfa_device_count_one?
+        session[:show_recovery_codes] = true
+        redirect_to recovery_multifactor_auth_path
+      else
+        redirect_to @continue_path
+        session.delete("mfa_redirect_uri")
+      end
     end
   end
 
   def update
+    session[:level] = level_param
+    @user = current_user
+
+    setup_mfa_authentication
+    setup_webauthn_authentication
+
+    create_new_mfa_expiry
+
+    render template: "multifactor_auths/mfa_prompt"
+  end
+
+  def mfa_update
+    if current_user.ui_mfa_verified?(params[:otp])
+      update_level_and_redirect
+    else
+      redirect_to edit_settings_path, flash: { error: t("multifactor_auths.incorrect_otp") }
+    end
+  end
+
+  def webauthn_update
+    @user = current_user
+    unless @user.webauthn_enabled?
+      redirect_to edit_settings_path, flash: { error: t("multifactor_auths.require_webauthn_enabled") }
+      return
+    end
+
+    if webauthn_credential_verified?
+      update_level_and_redirect
+    else
+      redirect_to edit_settings_path, flash: { error: @webauthn_error }
+    end
+  end
+
+  def destroy
     if current_user.ui_mfa_verified?(otp_param)
-      handle_new_level_param
+      flash[:success] = t(".success")
+      current_user.disable_totp!
       redirect_to session.fetch("mfa_redirect_uri", edit_settings_path)
       session.delete("mfa_redirect_uri")
     else
       flash[:error] = t("multifactor_auths.incorrect_otp")
       redirect_to edit_settings_path
     end
+  end
+
+  def recovery
+    if session[:show_recovery_codes].nil?
+      redirect_to edit_settings_path
+      flash[:error] = t(".already_generated")
+      return
+    end
+    @continue_path = session.fetch("mfa_redirect_uri", edit_settings_path)
+    session.delete("mfa_redirect_uri")
+  ensure
+    session.delete(:show_recovery_codes)
   end
 
   private
@@ -51,15 +110,23 @@ class MultifactorAuthsController < ApplicationController
     request.host || "rubygems.org"
   end
 
-  def require_mfa_disabled
-    return unless current_user.mfa_enabled?
-    flash[:error] = t("multifactor_auths.require_mfa_disabled")
+  def require_totp_disabled
+    return if current_user.totp_disabled?
+    flash[:error] = t("multifactor_auths.require_totp_disabled")
     redirect_to edit_settings_path
   end
 
   def require_mfa_enabled
     return if current_user.mfa_enabled?
     flash[:error] = t("multifactor_auths.require_mfa_enabled")
+    redirect_to edit_settings_path
+  end
+
+  def require_totp_enabled
+    return if current_user.totp_enabled?
+
+    flash[:error] = t("multifactor_auths.require_totp_enabled")
+    delete_mfa_level_update_session_variables
     redirect_to edit_settings_path
   end
 
@@ -71,18 +138,51 @@ class MultifactorAuthsController < ApplicationController
     end
   end
 
+  def setup_mfa_authentication
+    return if current_user.totp_disabled?
+    @form_mfa_url = mfa_update_multifactor_auth_url(token: current_user.confirmation_token)
+  end
+
+  def setup_webauthn_authentication
+    return if current_user.webauthn_disabled?
+
+    @webauthn_verification_url = webauthn_update_multifactor_auth_url(token: current_user.confirmation_token)
+
+    @webauthn_options = current_user.webauthn_options_for_get
+
+    session[:webauthn_authentication] = {
+      "challenge" => @webauthn_options.challenge
+    }
+  end
+
+  def update_level_and_redirect
+    handle_new_level_param
+    redirect_to session.fetch("mfa_redirect_uri", edit_settings_path)
+    session.delete(:mfa_redirect_uri)
+  end
+
   # rubocop:disable Rails/ActionControllerFlashBeforeRender
   def handle_new_level_param
-    case level_param
-    when "disabled"
-      flash[:success] = t("multifactor_auths.destroy.success")
-      current_user.disable_totp!
-    when "ui_only"
-      flash[:error] = t("multifactor_auths.ui_only_warning")
+    case session[:level]
+    when "ui_and_api", "ui_and_gem_signin"
+      flash[:success] = t("multifactor_auths.update.success")
+      current_user.update!(mfa_level: session[:level])
     else
-      flash[:error] = t(".success")
-      current_user.update!(mfa_level: level_param)
+      flash[:error] = t("multifactor_auths.update.invalid_level")
     end
   end
   # rubocop:enable Rails/ActionControllerFlashBeforeRender
+
+  def verify_session_expiration
+    return if session_active?
+
+    delete_mfa_level_update_session_variables
+    redirect_to edit_settings_path, flash: { error: t("multifactor_auths.session_expired") }
+  end
+
+  def delete_mfa_level_update_session_variables
+    session.delete(:level)
+    session.delete(:webauthn_authentication)
+    delete_mfa_expiry_session
+  end
 end
